rails 7.0.0 compatibility
It seems that rails 7 executes middleware.swap operations in a different
way, leading to failed middleware stack builds when other gems rely on a
specific middleware being there.

We avoid the problem by changing swap into a combination of insert and
delete. This works, because delete operations are executed after all
other operations.